[UPD+REF] Started to change unnecessary API Routes with Server Components (future default)
