+++ app/(routes)/(pages)/home/components/map-section/Floor.tsx
