--- app/(routes)/(pages)/login/error/page.tsx
@@ -1,11 +1,0 @@
-const Error: React.FC = () => {
-  return (
-    <main>
-      <div>
-        Eddu gayming
-      </div>
-    </main>
-  )
-}
-
-export default Error;
