+++ app/error.tsx
@@ -1,0 +1,11 @@
+"use client"
+
+const Error = () => {
+  return (
+    <div>
+      Error
+    </div>
+  );
+}
+
+export default Error;
