--- components/custom/auth/session-provider.tsx
@@ -1,3 +1,0 @@
-"use client"
-import { SessionProvider } from "next-auth/react"
-export default SessionProvider;
